+++ src/Block/System/Config/Summary.php
@@ -19,17 +19,25 @@ class Summary extends \Magento\Config\Block\System\Config\Form\Field
 {
     protected $_buttonLabel = '';
 
-    /** @var */
+    /** @var \Verifone\Payment\Model\Client\Config */
     protected $_config;
+
+    /**
+     * @var \Verifone\Payment\Helper\Urls
+     */
+    protected $_urls;
 
     public function __construct(
         \Magento\Backend\Block\Template\Context $context, 
         Config $config,
+        \Verifone\Payment\Helper\Urls $urls,
+
         array $data = [])
     {
         parent::__construct($context, $data);
         
         $this->_config = $config;
+        $this->_urls = $urls;
 
         $this->_buttonLabel = __('Display configuration summary');
     }
@@ -169,7 +177,7 @@ class Summary extends \Magento\Config\Block\System\Config\Form\Field
         return $data;
     }
 
-    public function getConfigurationDataForDisplay()
+    public function getConfigurationDataForDisplay2()
     {
         $data = $this->getConfigurationData();
 
@@ -258,6 +266,176 @@ class Summary extends \Magento\Config\Block\System\Config\Form\Field
 
         return $display;
 
+    }
+
+    public function getConfigurationDataForDisplay()
+    {
+
+        $helper = $this->_config;
+
+        if(!empty($this->_request->getParam('website'))){
+            $code = $this->_storeManager->getWebsite($this->_request->getParam('website'))->getCode();
+        } else {
+            $code = null;
+        }
+
+        /** Data for display */
+        $display = array();
+
+        $display['isLiveMode'] = array(
+            'label' => __('Mode'),
+            'value' => $this->_getScopeConfig(Path::XML_PATH_IS_LIVE_MODE, $code) ? __('Production') : __('Test'),
+            'has_desc' => false,
+            'has_desc_class' => false
+        );
+
+        $display['merchantCode'] = array(
+            'label' => __('Verifone Payment merchant agreement code'),
+            'value' => $helper->getMerchantAgreement($code),
+            'has_desc' => false,
+            'has_desc_class' => false
+        );
+        if ($helper->getMerchantAgreement($code) === $helper->getMerchantAgreementDefault($code)) {
+            $display['merchantCode']['desc'] = __('Default test merchant agreement uses');
+            $display['merchantCode']['desc_class'] = 'info';
+            $display['merchantCode']['has_desc'] = true;
+            $display['merchantCode']['has_desc_class'] = true;
+        }
+
+        $display['delayedUrl'] = array(
+            'label' => __('Delayed success url'),
+            'value' => $this->_urls->getSuccessDelayedUrl(),
+            'desc' => __('This is the url that you need to copy to payment provider settings in their portal.'),
+            'desc_class' => 'success',
+            'has_desc' => true,
+            'has_desc_class' => true
+        );
+
+        $display['keyHandlingMode'] = array(
+            'label' => __('Key handling mode'),
+            'value' => $helper->isKeySimpleMode($code) ? __('Automatic (Simple)') : __('Manual (Advanced)'),
+            'has_desc' => false,
+            'has_desc_class' => false
+        );
+
+        $display['paymentServiceKey'] = array(
+            'label' => __('Path and filename of Verifone Payment public key file'),
+            'value' => $helper->getPaymentPublicKeyPath($code),
+            'has_desc' => false,
+            'has_desc_class' => false
+        );
+
+        if (file_exists($helper->getPaymentPublicKeyPath($code))) {
+            $display['paymentServiceKey']['desc'] = __('Key file is available');
+            $display['paymentServiceKey']['desc_class'] = 'success';
+            $display['paymentServiceKey']['has_desc'] = true;
+            $display['paymentServiceKey']['has_desc_class'] = true;
+        } else {
+            $display['paymentServiceKey']['desc'] = __('Problem with load key file. Please contact with customer service');
+            $display['paymentServiceKey']['desc_class'] = 'success';
+            $display['paymentServiceKey']['has_desc'] = true;
+            $display['paymentServiceKey']['has_desc_class'] = true;
+        }
+
+        if ($helper->isKeyAdvancedMode($code)) {
+
+            $path = $helper->getKeysDirectory($code);
+
+            $display['directory'] = array(
+                'label' => __('Directory for store keys'),
+                'value' => $path,
+                'has_desc' => false,
+                'has_desc_class' => false
+            );
+            if (file_exists($path) && is_writable($path)) {
+                $display['directory']['desc'] = __('Directory configured properly');
+                $display['directory']['desc_class'] = 'success';
+                $display['directory']['has_desc'] = true;
+                $display['directory']['has_desc_class'] = true;
+            } else {
+                $display['directory']['desc'] = __('Problem with directory configuration. Please check configuration and save.');
+                $display['directory']['desc_class'] = 'error';
+                $display['directory']['has_desc'] = true;
+                $display['directory']['has_desc_class'] = true;
+            }
+
+            if($helper->getMerchantAgreementDefault($code) !== $helper->getMerchantAgreement($code)) {
+                $display['shopPrivateKey'] = array(
+                    'label' => __('Path and filename of shop private key file'),
+                    'value' => $helper->getShopPrivateKeyPath($code),
+                    'has_desc' => false,
+                    'has_desc_class' => false
+                );
+
+                if (file_exists($display['shopPrivateKey']['value']) && !empty($helper->getShopPrivateKeyFileName($code))) {
+                    $display['shopPrivateKey']['desc'] = __('Key file is available');
+                    $display['shopPrivateKey']['desc_class'] = 'success';
+                    $display['shopPrivateKey']['has_desc'] = true;
+                    $display['shopPrivateKey']['has_desc_class'] = true;
+                } else {
+                    $display['shopPrivateKey']['desc'] = __('Key file is not available');
+                    $display['shopPrivateKey']['desc_class'] = 'error';
+                    $display['shopPrivateKey']['has_desc'] = true;
+                    $display['shopPrivateKey']['has_desc_class'] = true;
+                }
+            } else {
+                $display['shopPrivateKey'] = array(
+                    'label' => __('Path and filename of shop private key file'),
+                    'value' => '',
+                    'has_desc' => true,
+                    'has_desc_class' => true,
+                    'desc' => __('Default key file is used'),
+                    'desc_class' => 'info'
+                );
+            }
+        } else {
+
+            $display['shopPrivateKey'] = array(
+                'label' => __('Path and filename of shop private key file'),
+                'value' => 'Key file stored in database',
+                'has_desc' => false,
+                'has_desc_class' => false
+            );
+
+            if ($helper->getShopPrivateKey($code) !== null && $helper->getShopPrivateKey($code) !== $helper->getShopPrivateKeyDefault($code)) {
+                $display['shopPrivateKey']['desc'] = __('Key file is available');
+                $display['shopPrivateKey']['desc_class'] = 'success';
+                $display['shopPrivateKey']['has_desc'] = true;
+                $display['shopPrivateKey']['has_desc_class'] = true;
+            } elseif(!$helper->isLiveMode($code) && $helper->getMerchantAgreement($code) === $helper->getMerchantAgreementDefault($code)) {
+                $display['shopPrivateKey']['desc'] = __('Default key file is used');
+                $display['shopPrivateKey']['desc_class'] = 'info';
+                $display['shopPrivateKey']['has_desc'] = true;
+                $display['shopPrivateKey']['has_desc_class'] = true;
+            } else {
+                $display['shopPrivateKey']['desc'] = __('Problem with fetch shop private key file. Please check configuration and/or generate key');
+                $display['shopPrivateKey']['desc_class'] = 'error';
+                $display['shopPrivateKey']['has_desc'] = true;
+                $display['shopPrivateKey']['has_desc_class'] = true;
+            }
+        }
+
+        if ($helper->isKeySimpleMode($code) && $helper->getShopPublicKey($code) === null && $helper->getMerchantAgreement($code) !== $helper->getMerchantAgreementDefault($code)) {
+            $display['shopPublicKeyContent'] = array(
+                'label' => __('Public key file'),
+                'value' => '',
+                'has_desc' => true,
+                'has_desc_class' => true,
+                'desc' => __('Problem with fetch shop public key file. Please check configuration and/or generate key'),
+                'desc_class' => 'error'
+            );
+        } elseif ($helper->getShopPublicKey($code) !== null) {
+            $display['shopPublicKeyContent'] = array(
+                'label' => __('Public key file'),
+                'value' => $helper->getShopPublicKey($code),
+                'has_desc' => true,
+                'has_desc_class' => true,
+                'desc' => __('Please, copy this key to payment operator configuration settings, otherwise, the payment will be broken.'),
+                'desc_class' => 'success'
+            );
+        }
+
+        return $display;
     }
 
     protected function _getScopeConfig($path, $websiteCode)