+++ src/Model/Client/Config.php
@@ -14,6 +14,7 @@ namespace Verifone\Payment\Model\Client;
 
 use Magento\Store\Model\ScopeInterface;
 use Verifone\Payment\Helper\Path as Path;
+use Verifone\Payment\Model\Config\Source\HandlingMode;
 
 class Config implements ConfigInterface
 {
@@ -56,6 +57,11 @@ class Config implements ConfigInterface
     protected $_helper;
 
     /**
+     * @var \Verifone\Payment\Helper\Keys
+     */
+    protected $_keysHelper;
+
+    /**
      * Config constructor.
      *
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
@@ -63,13 +69,15 @@ class Config implements ConfigInterface
      * @param \Magento\Framework\App\ProductMetadataInterface $productMetadata
      * @param \Magento\Framework\Module\Dir\Reader $reader
      * @param \Verifone\Payment\Helper\Payment $helper
+     * @param \Verifone\Payment\Helper\Keys $keysHelper
      */
     public function __construct(
         \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
         \Magento\Store\Model\StoreManagerInterface $storeManager,
         \Magento\Framework\App\ProductMetadataInterface $productMetadata,
         \Magento\Framework\Module\Dir\Reader $reader,
-        \Verifone\Payment\Helper\Payment $helper
+        \Verifone\Payment\Helper\Payment $helper,
+        \Verifone\Payment\Helper\Keys $keysHelper
     )
     {
         $this->_scopeConfig = $scopeConfig;
@@ -77,6 +85,7 @@ class Config implements ConfigInterface
         $this->_moduleReader = $reader;
         $this->_storeManager = $storeManager;
         $this->_helper = $helper;
+        $this->_keysHelper = $keysHelper;
     }
 
     /**
@@ -106,8 +115,8 @@ class Config implements ConfigInterface
     {
         if (!$this->isConfigSet()) {
 
-            $privateKeyPath = $this->getShopKeyFile();
-            $publicKeyPath = $this->getPaypageKeyFile();
+            $privateKeyPath = $this->getShopPrivateKeyFile();
+            $publicKeyPath = $this->getPaymentPublicKeyFile();
             $merchant = $this->getMerchantAgreement();
             $software = 'Magento';
             $softwareVersion = $this->_productMetadata->getVersion();
@@ -161,137 +170,186 @@ class Config implements ConfigInterface
         return $this->_getScopeConfig(Path::XML_PATH_MERCHANT_CODE_DEFAULT, $code);
     }
 
-    ///
-    /// Shop Private Key
-    ///
-    public function getShopKeyFile($code = null)
+
+    public function getKeyMode($code = null)
     {
-        $customKey = $this->getShopKeyFileCustom($code);
-
-        if (null !== $customKey && file_exists($customKey)) {
-            return $customKey;
-        }
-
-        return $this->getShopKeyFileDefault();
+        return $this->_getScopeConfig(Path::XML_PATH_KEY_MODE, $code);
     }
 
-    public function getShopKeyFileCustom($code = null)
+    public function isKeySimpleMode($code = null)
+    {
+        return (int)$this->getKeyMode($code) === HandlingMode::SIMPLE_MODE;
+    }
+
+    public function isKeyAdvancedMode($code = null)
+    {
+        return (int)$this->getKeyMode($code) === HandlingMode::ADVANCED_MODE;
+    }
+
+
+    /** KEYS FILES */
+    public function isLiveMode($code = null)
+    {
+        return $this->_getScopeConfig(Path::XML_PATH_IS_LIVE_MODE, $code);
+    }
+
+    public function getKeysDirectory($code = null)
+    {
+        return $this->_getScopeConfig(Path::XML_PATH_KEY_DIRECTORY, $code);
+    }
+
+    public function getShopPrivateKeyFileName($code = null)
+    {
+        if ($this->isLiveMode($code)) {
+            return $this->_getScopeConfig(Path::XML_PATH_KEY_SHOP, $code);
+        }
+
+        return $this->_getScopeConfig(Path::XML_PATH_KEY_SHOP_TEST, $code);
+    }
+
+    public function getLiveShopPrivateKeyPath($code = null)
     {
 
-        $key = $this->getKeyCustom(Path::XML_PATH_KEY_SHOP);
-
-        if (($key === null || !file_exists($key)) && $this->getMerchantAgreement($code) !== $this->getMerchantAgreementDefault($code)) {
-            $directory = $this->_getScopeConfig(Path::XML_PATH_KEY_DIRECTORY, $code);
-            $fileName = $this->getMerchantAgreement($code) . '-private.pem';
-
-            return $directory . DIRECTORY_SEPARATOR . $fileName;
-        } elseif ($key !== null && $this->getMerchantAgreement($code) === $this->getMerchantAgreementDefault($code)) {
+        if ($this->isKeySimpleMode($code)) {
             return null;
         }
 
-        return $key;
+        return $this->getKeysDirectory($code) . DIRECTORY_SEPARATOR . $this->_getScopeConfig(Path::XML_PATH_KEY_SHOP, $code);
+
     }
 
-    public function getShopKeyFileDefault($code = null)
+    public function getTestShopPrivateKeyPath($code = null)
     {
-        return $this->getKeyDefault(Path::XML_PATH_KEY_SHOP_DEFAULT);
-    }
 
-    ///
-    /// Shop Public Key
-    ///
-    public function getPublicShopKeyFile($code = null)
-    {
-        $customKey = $this->getPublicShopKeyFileCustom($code);
-
-        if (null !== $customKey && file_exists($customKey)) {
-            return $customKey;
+        if ($this->isKeySimpleMode($code)) {
+            return null;
         }
 
-        return $this->getPublicShopKeyFileDefault();
+        return $this->getKeysDirectory($code) . DIRECTORY_SEPARATOR . $this->_getScopeConfig(Path::XML_PATH_KEY_SHOP_TEST, $code);
     }
 
-    public function getPublicShopKeyFileCustom($code = null)
+    public function getShopPrivateKeyPath($code = null)
     {
-        if ($this->getMerchantAgreement($code) !== $this->getMerchantAgreementDefault($code)) {
-            $directory = $this->_getScopeConfig(Path::XML_PATH_KEY_DIRECTORY, $code);
-
-            $fileName = $this->getMerchantAgreement($code) . '-public.pem';
-            return $directory . DIRECTORY_SEPARATOR . $fileName;
+        if ($this->isLiveMode($code)) {
+            return $this->getLiveShopPrivateKeyPath($code);
         }
 
-        return null;
-
+        return $this->getTestShopPrivateKeyPath($code);
     }
 
-    public function getPublicShopKeyFileDefault($code = null)
+    public function getShopPrivateKeyFile($code = null)
     {
-        return $this->getKeyDefault(Path::XML_PATH_KEY_SHOP_PUBLIC_DEFAULT);
+        return $this->getShopPrivateKey();
     }
 
-    public function getPublicShopKeyContent($code = null)
+    public function getShopPrivateKey($code = null)
     {
-        return \file_get_contents($this->getPublicShopKeyFile($code));
-    }
 
-    ///
-    /// Payment Service Public Key
-    ///
-    public function getPaypageKeyFile($code = null)
-    {
-        $customKey = $this->getPaypageKeyFileCustom($code, $code);
+        // If TEST mode is set
+        if (!$this->isLiveMode($code)) {
 
-        if (null !== $customKey && file_exists($customKey)) {
-            return $customKey;
-        }
-
-        if ($this->_getScopeConfig(Path::XML_PATH_IS_LIVE_MODE, $code)) {
-            return $this->getPaypageKeyFileLiveDefault($code);
-        }
-
-        return $this->getPaypageKeyFileDefault();
-    }
-
-    public function getPaypageKeyFileCustom($code = null)
-    {
-        return $this->getKeyCustom(PATH::XML_PATH_KEY_VERIFONE);
-    }
-
-    public function getPaypageKeyFileLiveDefault($code = null)
-    {
-        return $this->getKeyDefault(Path::XML_PATH_KEY_VERIFONE_LIVE_DEFAULT);
-    }
-
-    public function getPaypageKeyFileDefault($code = null)
-    {
-        return $this->getKeyDefault(Path::XML_PATH_KEY_VERIFONE_TEST_DEFAULT);
-    }
-
-    public function getKeyCustom($keyConfigurationPath, $code = null)
-    {
-        $directory = $this->_getScopeConfig(Path::XML_PATH_KEY_DIRECTORY, $code);
-
-        $fileName = '';
-
-        if ($this->_getScopeConfig(Path::XML_PATH_IS_LIVE_MODE, $code)) {
-            $fileName = $this->_getScopeConfig($keyConfigurationPath, $code);
-        } else {
-            if (!empty($this->_getScopeConfig($keyConfigurationPath . '_test', $code))) {
-                $fileName = $this->_getScopeConfig($keyConfigurationPath . '_test', $code);
+            if ($this->getMerchantAgreement($code) === $this->getMerchantAgreementDefault($code)) {
+                // If DEFAULT test merchant is set, return default key
+                return $this->getShopPrivateKeyDefault($code);
             }
+
+            if ($this->isKeySimpleMode($code)) {
+                // If CUSTOM test merchant is set, and SIMPLE mode is set, return generated key stored in DB
+                return $this->_keysHelper->getTestPrivateKey($code);
+            }
+
+            $path = $this->getTestShopPrivateKeyPath($code);
+            if (file_exists($path)) {
+                // If CUSTOM test merchant is set, and ADVANCED mode is set, return old key stored in files
+                return file_get_contents($path);
+            }
+
+            // return default key file
+            return $this->getShopPrivateKeyDefault($code);
         }
 
-        if (empty($fileName)) {
+        // If LIVE mode is set
+
+        if ($this->isKeySimpleMode($code)) {
+            // If CUSTOM test merchant is set, and SIMPLE mode is set, return generated key stored in DB
+            return $this->_keysHelper->getLivePrivateKey($code);
+        }
+
+        $path = $this->getLiveShopPrivateKeyPath($code);
+        if (file_exists($path)) {
+            // If CUSTOM test merchant is set, and ADVANCED mode is set, return old key stored in files
+            return file_get_contents($path);
+        }
+
+        // return nothing
+        return null;
+    }
+
+    public function getShopPrivateKeyPathDefault($code = null)
+    {
+        return $this->_moduleReader->getModuleDir('', 'Verifone_Payment') . DIRECTORY_SEPARATOR . 'keys' . DIRECTORY_SEPARATOR . $this->_getScopeConfig(Path::XML_PATH_KEY_SHOP_DEFAULT, $code);
+    }
+
+    public function getShopPrivateKeyDefault($code = null)
+    {
+        return file_get_contents($this->getShopPrivateKeyPathDefault($code));
+    }
+
+    public function getShopPublicKey($code = null)
+    {
+        // If TEST mode is set
+        if (!$this->isLiveMode($code)) {
+
+            if ($this->getMerchantAgreement($code) === $this->getMerchantAgreementDefault($code)) {
+                // If DEFAULT test merchant is set, return default key
+                // For default simple key is not require to configure in payment service.
+                return null;
+            }
+
+            if ($this->isKeySimpleMode($code)) {
+                // If CUSTOM test merchant is set, and SIMPLE mode is set, return generated key stored in DB
+                return $this->_keysHelper->getTestPublicKey($code);
+            }
+
+
+            // If CUSTOM test merchant is set, and ADVANCED mode is set, return old key stored in files
+            // This is not required, because if this mode is set, it means that payment service is configured,
+            // and it does not require to configure again.
             return null;
         }
 
-        return $directory . DIRECTORY_SEPARATOR . $fileName;
+        // If LIVE mode is set
+        if ($this->isKeySimpleMode($code)) {
+            // If CUSTOM test merchant is set, and SIMPLE mode is set, return generated key stored in DB
+            return $this->_keysHelper->getLivePublicKey($code);
+        }
+
+        // When advanced mode is set, or is default merchant agreement then in not require simple key to display.
+        return null;
     }
 
-    public function getKeyDefault($keyConfigurationPath, $code = null)
+    public function getShopPublicKeyPathDefault($code = null)
     {
-        $moduleDir = $this->_moduleReader->getModuleDir('', 'Verifone_Payment');
-        return $moduleDir . DIRECTORY_SEPARATOR . 'keys' . DIRECTORY_SEPARATOR . $this->_getScopeConfig($keyConfigurationPath, $code);
+        return $this->_moduleReader->getModuleDir('', 'Verifone_Payment') . DIRECTORY_SEPARATOR . 'keys' . DIRECTORY_SEPARATOR . $this->_getScopeConfig(Path::XML_PATH_KEY_SHOP_PUBLIC_DEFAULT, $code);
+    }
+
+    public function getShopPublicKeyDefault($code = null)
+    {
+        return file_get_contents($this->getShopPublicKeyPathDefault($code));
+    }
+
+    public function getPaymentPublicKeyPath($code = null)
+    {
+        if ($this->isLiveMode($code)) {
+            return $this->_moduleReader->getModuleDir('', 'Verifone_Payment') . DIRECTORY_SEPARATOR . 'keys' . DIRECTORY_SEPARATOR . $this->_getScopeConfig(Path::XML_PATH_KEY_VERIFONE_LIVE_DEFAULT, $code);
+        }
+
+        return $this->_moduleReader->getModuleDir('', 'Verifone_Payment') . DIRECTORY_SEPARATOR . 'keys' . DIRECTORY_SEPARATOR . $this->_getScopeConfig(Path::XML_PATH_KEY_VERIFONE_TEST_DEFAULT, $code);
+    }
+
+    public function getPaymentPublicKeyFile($code = null)
+    {
+        return file_get_contents($this->getPaymentPublicKeyPath($code));
     }
 
     protected function _getScopeConfig($path, $websiteCode)
@@ -303,14 +361,9 @@ class Config implements ConfigInterface
         return $this->_scopeConfig->getValue($path);
     }
 
-    /**
-     * @param $filepath
-     *
-     * @return string
-     */
-    public function getFileContent($filepath)
+    public function getFileContent($string)
     {
-        return file_get_contents($filepath);
+        return $string;
     }
 
     /**